requirements.txt
openai
requests
pandas
transformers
faiss-cpu
python-dotenv